pr comments improvements
