+++ hl7_transformer/hl7_transformer/datetime_transformer.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from typing import Optional
 
 
 def transform_datetime(date_time: str) -> str:
@@ -11,14 +10,3 @@ def transform_datetime(date_time: str) -> str:
     except ValueError:
         dt = datetime.strptime(date_time, "%Y-%m-%d %H:%M:%S")
         return dt.strftime(required_format)
-
-def transform_date_of_death(date_of_death: Optional[str]) -> str:
-    if not date_of_death or not date_of_death.strip():
-        return ""
-    
-    normalized_value = date_of_death.strip().upper()
-    
-    if normalized_value == "RESURREC":
-        return '""'
-    
-    return date_of_death.strip()
